A comparison script between the package dbs
This is a simple comparison between the packages Puppet is managing versus those
that are installed according to pacman.

The eventual goal is being able to purge all packages that aren't installed by
Puppet from the system. Right now, that would be REALLY bad, seeing as it would
uninstall things like glibc and the kernel. :P